rtah_apis/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
djangorestframework==3.12.4
djangorestframework-api-key==2.1.0
idna==3.2
Markdown==3.3.4
mysqlclient==2.0.3
Pillow==8.3.2
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
urllib3==1.26.6
